requirements.txt
aiogram==2.25.1 ; python_version >= "3.7" and python_version < "4.0"
aiohttp==3.8.5 ; python_version >= "3.7" and python_version < "4.0"
aiosignal==1.3.1 ; python_version >= "3.7" and python_version < "4.0"
async-timeout==4.0.2 ; python_version >= "3.7" and python_version < "4.0"
asyncpg==0.26.0 ; python_version >= "3.7" and python_version < "4.0"
asynctest==0.13.0 ; python_version >= "3.7" and python_version < "3.8"
attrs==23.1.0 ; python_version >= "3.7" and python_version < "4.0"
babel==2.9.1 ; python_version >= "3.7" and python_version < "4.0"
certifi==2023.7.22 ; python_version >= "3.7" and python_version < "4.0"
charset-normalizer==3.1.0 ; python_version >= "3.7" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.7" and python_version < "4.0" and sys_platform == "win32"
frozenlist==1.3.3 ; python_version >= "3.7" and python_version < "4.0"
idna==3.4 ; python_version >= "3.7" and python_version < "4.0"
importlib-metadata==4.2.0 ; python_version >= "3.7" and python_version < "3.8"
loguru==0.6.0 ; python_version >= "3.7" and python_version < "4.0"
magic-filter==1.0.9 ; python_version >= "3.7" and python_version < "4.0"
multidict==6.0.4 ; python_version >= "3.7" and python_version < "4.0"
python-dotenv==0.21.1 ; python_version >= "3.7" and python_version < "4.0"
pytz==2023.3 ; python_version >= "3.7" and python_version < "4.0"
redis==4.5.5 ; python_version >= "3.7" and python_version < "4.0"
typing-extensions==4.6.3 ; python_version >= "3.7" and python_version < "3.8"
#uvloop==0.17.0 ; python_version >= "3.7" and python_version < "4.0"
win32-setctime==1.1.0 ; python_version >= "3.7" and python_version < "4.0" and sys_platform == "win32"
yarl==1.9.2 ; python_version >= "3.7" and python_version < "4.0"
zipp==3.15.0 ; python_version >= "3.7" and python_version < "3.8"
httpx
pydantic